tests/wpt/tests/scroll-animations/scroll-timelines/duration.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ScrollTimeline.duration</title>
<link rel="help" href="https://drafts.csswg.org/web-animations-2/#dom-animationtimeline-duration">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
'use strict';

test(t => {
  const timeline = new ScrollTimeline();
  assert_equals(timeline.duration.toString(), '100%');
}, 'The duration of a scroll timeline is 100%');

</script>
